+++ servers/factorio/deployment.py
@@ -1,6 +1,6 @@
 import aiohttp
 from core.util import util, tasks, io, pack, aggtrf
-from core.msg import msgabc, msgext, msgftr
+from core.msg import msgabc, msgext, msgftr, msglog
 from core.context import contextsvc
 from core.http import httpabc, httprsc, httpext, httpsubs
 from core.proc import proch, jobh
@@ -69,11 +69,11 @@ class Deployment:
 
     def resources(self, resource: httpabc.Resource):
         archive_selector = httpsubs.Selector(
-            msg_filter=msgftr.NameIs(msgext.LoggingPublisher.INFO),
+            msg_filter=msgftr.NameIs(msglog.LoggingPublisher.INFO),
             completed_filter=msgftr.DataEquals('END Archive Directory'),
             aggregator=aggtrf.StrJoin('\n'))
         unpacker_selector = httpsubs.Selector(
-            msg_filter=msgftr.NameIs(msgext.LoggingPublisher.INFO),
+            msg_filter=msgftr.NameIs(msglog.LoggingPublisher.INFO),
             completed_filter=msgftr.DataEquals('END Unpack Directory'),
             aggregator=aggtrf.StrJoin('\n'))
         httprsc.ResourceBuilder(resource) \
