+++ servers/serverlink/server.py
@@ -1,9 +1,9 @@
 from core.context import contextsvc, contextext
 from core.http import httpabc, httpsubs, httprsc, httpext
-from core.msg import msgext, msgftr, msgtrf
+from core.msg import msgftr, msgtrf, msglog
 from core.proc import proch, prcext
 from core.system import svrabc, svrsvc, svrext
-from core.util import io
+from core.util import util, io
 
 
 class Server(svrabc.Server):
@@ -21,7 +21,7 @@ class Server(svrabc.Server):
     async def initialise(self):
         await self._server_process_factory.initialise()
         self._context.register(prcext.ServerStateSubscriber(self._context))
-        self._context.register(msgext.LogfileSubscriber(
+        self._context.register(msglog.LogfileSubscriber(
             self._log_file,
             msgftr.Or(proch.ServerProcess.FILTER_STDOUT_LINE, proch.ServerProcess.FILTER_STDERR_LINE),
             transformer=msgtrf.GetData()))
@@ -53,20 +53,26 @@ class _ServerProcessFactory:
         self._context = context
         self._config = config
         self._clientfile = clientfile
-        self._executable = '/usr/local/bin/node'
-        self._exe_exists = False
+        self._executable = None
+        self._use_source = False
 
     async def initialise(self):
-        alternates = [self._context.config('home') + '/serverlink', '/usr/local/bin/serverlink']
-        for executable in alternates:
-            if await io.file_exists(executable):
-                self._executable = executable
-                self._exe_exists = True
-                return
+        self._executable = self._context.config('home') + '/serverlink'
+        if await io.file_exists(self._executable):
+            return
+        env_path = util.get('PATH', self._context.config('env'))
+        self._executable = await io.find_in_env_path(env_path, 'serverlink')
+        if self._executable:
+            return
+        self._executable = await io.find_in_env_path(env_path, 'node')
+        if self._executable:
+            self._use_source = True
+            return
+        raise Exception('Unable to find a ServerLink executable.')
 
     def build(self) -> proch.ServerProcess:
         server_process = proch.ServerProcess(self._context, self._executable)
-        if not self._exe_exists:
+        if self._use_source:
             server_process.append_arg(self._context.config('home') + '/index.js')
         server_process.append_arg(self._config)
         server_process.append_arg(self._clientfile)
